(fleet) re-organize the packages package
This PR is a cleanup of the packages package to prepare a migration to a proper strict `prerm`+`postrm` model.

For now it's mostly moving the systemd, os package manager and apminject installer to their own subpackages.
