+++ pkg/fleet/installer/packages/apminject/app_armor.go
@@ -5,7 +5,7 @@
 
 //go:build !windows
 
-package packages
+package apminject
 
 import (
 	"bufio"
