+++ pkg/fleet/installer/packages/apminject/app_armor_test.go
@@ -5,14 +5,15 @@
 
 //go:build !windows
 
-package packages
+package apminject
 
 import (
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 	"os"
 	"path/filepath"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 )
 
 func TestFindAndReplaceFile(t *testing.T) {
