+++ pkg/fleet/installer/packages/apminject/file.go
@@ -5,7 +5,7 @@
 
 //go:build !windows
 
-package packages
+package apminject
 
 import (
 	"bytes"
