+++ pkg/fleet/installer/packages/apminject/file_test.go
@@ -5,7 +5,7 @@
 
 //go:build !windows
 
-package packages
+package apminject
 
 import (
 	"context"
